pyproject.toml
[project]
name = "rappi-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dash>=4.1.0",
    "dash-bootstrap-components>=2.0.4",
    "google-adk>=1.31.0",
    "litellm>=1.83.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.2",
    "plotly>=6.7.0",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "pytest>=9.0.3",
    "ruff>=0.15.11",
]
